+++ routes/web.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 use Inertia\Response;
 
-Route::get("/signup", fn(): Response => inertia::render("Auth/Signup"));
+Route::get("/signup", fn(): Response => inertia::render("Auth/Register"));
 Route::get("/login", [LoginController::class, "create"])->name("login");
 Route::post("/login", [LoginController::class, "login"])->name("login");
 Route::post("/register", [RegisterController::class, "store"])->name("register");
